+++ GameEngine/Tools/AssetsEditor/Convertor/PhysicsLoader.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "CookingTask.hpp"
+
+namespace AGE
+{
+	struct PhysicsLoader
+	{
+		static bool save(std::shared_ptr<CookingTask> cookingTask);
+
+		static bool load(std::shared_ptr<CookingTask> cookingTask);
+	};
+}
